Dev hamdi

+++ database/migrations/2021_11_11_140555_drop_foreign_keys_in_products_table.php
@@ -13,12 +13,15 @@ class DropForeignKeysInProductsTable extends Migration
      */
     public function up()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropForeign(['stock_type_id']);
-            $table->dropColumn('stock_type_id');
-            $table->dropForeign(['unity_id']);
-            $table->dropColumn('unity_id');
-        });
+
+        if (Schema::hasColumn('products', 'stock_type_id')) {
+            Schema::table('products', function (Blueprint $table) {
+                // $table->dropForeign(['stock_type_id']);
+                // $table->dropColumn('stock_type_id');
+                $table->dropForeign(['unity_id']);
+                $table->dropColumn('unity_id');
+            });
+        }
     }
 
     /**
